Changed material ID to an uint

+++ includes/common/Vertex.h
@@ -11,7 +11,7 @@ struct Vertex {
 	glm::vec4 coords = glm::vec4(0.0f, 0.0f, 0.0f, 1.0f);
 	glm::vec3 normal = glm::vec3(1.0f, 1.0f, 1.0f);
 	glm::vec2 tex_coord = glm::vec2(0.0f, 0.0f);
-	GLfloat material_id = 0.0f; // isto e uma float por simplicidade, vao ser poucos vertices e vao, senao podia ser uint8 ou 16
+	GLuint material_id = 0;
 	GLfloat transf_id = 0.0f; // same
 
     friend std::ostream &operator<<(std::ostream &os, const Vertex &p) {
@@ -19,7 +19,7 @@ struct Vertex {
         return os;
     }
 
-	Vertex(float x, float y, float z, float normal_x, float normal_y, float normal_z, float tex_x, float tex_y, float _material_id)
+	Vertex(float x, float y, float z, float normal_x, float normal_y, float normal_z, float tex_x, float tex_y, GLuint _material_id)
 	: coords(x, y, z, 1.0f), normal(normal_x, normal_y, normal_z), tex_coord(tex_x, tex_y), material_id(_material_id) {}
     Vertex(float x, float y, float z) : coords(x, y, z, 1.0f) {}
     Vertex() = default;
